Add managed reactor-core and reactor-test to catalog
When switching microstream to the latest build for micronaut 4.0.0, it raised an issue

https://github.com/micronaut-projects/micronaut-microstream/pull/144

We could pull reactor-core from the core-bom, but as it didn't contain reactor-test we had to jump through hoops.

In https://github.com/micronaut-projects/micronaut-reactor/pull/248 we swiched to using the reactor bom instead of managed dependencies.

But as the reactor-bom hard-codes versions instead of setting them as properties we need to specify the version of reactor-core and reactor-test that we want to include.

The downside here is we will need to keep them in sync whenever the BOM version gets updated